CURB-3820 Renamed push opt in trigger related files
