+++ libraries/engage/push/action-creators/index.js
@@ -1,1 +1,1 @@
-export * from './optIn';
+export * from './optInTrigger';
