+++ libraries/engage/push/actions/index.js
@@ -1,1 +1,1 @@
-export * from './optIn';
+export * from './optInTrigger';
