+++ libraries/engage/push/actions/optInTrigger.js
@@ -5,4 +5,5 @@
  */
 export const showOptIn = () => async () => {
   // TODO implement logic
+  console.warn('SHOW_PUSH_OPT_IN');
 };
