+++ libraries/engage/push/selectors/index.js
@@ -1,1 +1,1 @@
-export * from './optIn';
+export * from './optInTrigger';
