+++ libraries/engage/push/selectors/optInTrigger.js
@@ -3,4 +3,4 @@
  * @param {Object} state The current state.
  * @returns {Object} The push opt in information.
  */
-export const getPushOptInState = state => state?.push?.optIn || {};
+export const getPushOptInTriggerState = state => state?.push?.optInTrigger || {};
