+++ themes/theme-gmd/pages/reducers.js
@@ -29,7 +29,7 @@ persistedReducers.set([
   'url',
   'user',
   'appRating',
-  'push.optIn',
+  'push.optInTrigger',
 ]);
 
 configuration.set(RESET_APP_REDUCERS, [
